moving newFormModal to a custom component so the index files are less crowded
